Execution engine: fix compile errors and add new graph module.
This commit fixes all compile errors from previous builds for both shell
and dash, and adds a new graph module to the dash package that represents
the new execution engine. It also adds a new runtime and client to the
runtime package that runs the client server handshake based on the new graph format.