execution: Add integration tests to check functionality of nodes.
This commit adds integration tests for dash, in `dash/tests`, in order
to check that the main functionality of cmdnodes, readnodes, and
writenodes correctly stream input and output in between each other, both
over pipes and TCP connections. This commit also fixes various bugs in
the execution engine necessary to get the tests working.